+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/),
 and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 
-## [2.4.0] - not released yet
+## [2.4.0] - 2021-06-11
 ### Changed
 - now `fpdf2` uses the newly supported `DCTDecode` image filter for JPEG images,
   instead of `FlateDecode` before, in order to improve the compression ratio without any image quality loss.
@@ -15,6 +15,7 @@ and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 - `FPDF.cell`: the `w` (width) parameter becomes optional, with a default value of `None`, meaning to generate a cell with the size of the text content provided
 - the `h` (height) parameter of the `cell`, `multi_cell` & `write` methods gets a default value change, `None`, meaning to use the current font size
 - removed the useless `w` & `h` parameters of the `FPDF.text_annotation()` method
+- the default `FPDF.producer` is now `PyFPDF/fpdf{FPDF_VERSION}`
 ### Added
 - new `FPDF.add_action()` method, documented in the [Annotations section](https://pyfpdf.github.io/fpdf2/Annotations.html)
 - `FPDF.cell`: new optional `markdown=True` parameter that enables basic Markdown-like styling: `**bold**, __italics__, --underlined--`
@@ -29,6 +30,10 @@ and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 - support for Code 39 barcodes in `fpdf.template`, using `type="C39"`
 ### Fixed
 - avoid an `Undefined font` error when using `write_html` with unicode bold or italics fonts
+### Deprecated
+- the `FPDF.set_doc_option()` method is deprecated in favour of just setting the `core_fonts_encoding` property
+  on an instance of `FPDF`
+- the `fpdf.SYSTEM_TTFONTS` configurable module constant is now ignored
 
 ## [2.3.5] - 2021-05-12
 ### Fixed